Fix CodeFactor issues - [SA1402]
