+++ Zilon.Core/Zilon.TextClient/GameScreen.cs
@@ -1,0 +1,10 @@
+﻿namespace Zilon.TextClient
+{
+    internal enum GameScreen
+    {
+        Undefinded,
+        GlobeSelection,
+        Main,
+        Scores
+    }
+}
